# stock-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stock Management API",
        "description": "API de gestion de stock multi-vendor et multi-entrepôt",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/dashboard/overview": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Vue d'ensemble du tableau de bord",
                "operationId": "651bc227644e3eaf947f107ac3d77adf",
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/sales-chart": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Graphique des ventes",
                "operationId": "9776c4d043b858f9d6db43fbd261cd65",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (week, month, year)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Données du graphique"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/top-products": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Produits les plus vendus",
                "operationId": "b45446e18fe41f2c035f6cccdd9a48a9",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre de produits",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top produits récupérés"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/recent-movements": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Mouvements de stock récents",
                "operationId": "c275a1f354c6a401513c8a01fc4d7cab",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre de mouvements",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mouvements récupérés"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/recent-orders": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Commandes récentes",
                "operationId": "d5fd3bcfe0ac30f88465d63ea8345b12",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre de commandes",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commandes récupérées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/stock-value": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Valeur du stock par entrepôt",
                "operationId": "bfbd042f972d85cd7a0f6ca674a29d30",
                "responses": {
                    "200": {
                        "description": "Valeurs récupérées"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/inventory-turnover": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Rotation des stocks",
                "operationId": "6280a6f04cdfe358b57e8fd26de3bc87",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (month, year)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rotation calculée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}